# backend/internal/store/internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql/driver"
	"fmt"
	"net/netip"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type RefreshStatus string

const (
	RefreshStatusActive  RefreshStatus = "active"
	RefreshStatusRotated RefreshStatus = "rotated"
	RefreshStatusRevoked RefreshStatus = "revoked"
)

func (e *RefreshStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = RefreshStatus(s)
	case string:
		*e = RefreshStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for RefreshStatus: %T", src)
	}
	return nil
}

type NullRefreshStatus struct {
	RefreshStatus RefreshStatus
	Valid         bool // Valid is true if RefreshStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullRefreshStatus) Scan(value interface{}) error {
	if value == nil {
		ns.RefreshStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.RefreshStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullRefreshStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.RefreshStatus), nil
}

type TokenPurpose string

const (
	TokenPurposeEmailVerification TokenPurpose = "email_verification"
	TokenPurposePasswordReset     TokenPurpose = "password_reset"
	TokenPurposeInvitation        TokenPurpose = "invitation"
)

func (e *TokenPurpose) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TokenPurpose(s)
	case string:
		*e = TokenPurpose(s)
	default:
		return fmt.Errorf("unsupported scan type for TokenPurpose: %T", src)
	}
	return nil
}

type NullTokenPurpose struct {
	TokenPurpose TokenPurpose
	Valid        bool // Valid is true if TokenPurpose is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTokenPurpose) Scan(value interface{}) error {
	if value == nil {
		ns.TokenPurpose, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TokenPurpose.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTokenPurpose) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TokenPurpose), nil
}

type UserStatus string

const (
	UserStatusPendingVerification UserStatus = "pending_verification"
	UserStatusActive              UserStatus = "active"
	UserStatusLocked              UserStatus = "locked"
	UserStatusDisabled            UserStatus = "disabled"
)

func (e *UserStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = UserStatus(s)
	case string:
		*e = UserStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for UserStatus: %T", src)
	}
	return nil
}

type NullUserStatus struct {
	UserStatus UserStatus
	Valid      bool // Valid is true if UserStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUserStatus) Scan(value interface{}) error {
	if value == nil {
		ns.UserStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.UserStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUserStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.UserStatus), nil
}

type Application struct {
	ID            uuid.UUID
	ClientID      string
	RedirectUri   string
	AllowedOrigin string
	Name          string
}

type AuditLog struct {
	ID           int64
	ActorUserID  pgtype.UUID
	Action       string
	ResourceType pgtype.Text
	ResourceID   pgtype.Text
	Ip           *netip.Addr
	UserAgent    pgtype.Text
	Metadata     []byte
	CreatedAt    pgtype.Timestamptz
}

type AuthorizationCode struct {
	ID            uuid.UUID
	UserID        uuid.UUID
	ApplicationID uuid.UUID
	CodeHash      []byte
	RedirectUri   string
	CodeChallenge string
	ExpiresAt     pgtype.Timestamptz
	UsedAt        pgtype.Timestamptz
}

type HubSession struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash []byte
	ExpiresAt pgtype.Timestamptz
	RevokedAt pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
	FamilyID  pgtype.UUID
}

type MfaChallenge struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	TokenHash    []byte
	CodeHash     []byte
	ExpiresAt    pgtype.Timestamptz
	AttemptsLeft int32
	LastSentAt   pgtype.Timestamptz
	UsedAt       pgtype.Timestamptz
	CreatedAt    pgtype.Timestamptz
}

type RecoveryCode struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	CodeHash  string
	UsedAt    pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
}

type RefreshToken struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	TokenHash  []byte
	FamilyID   uuid.UUID
	ParentID   pgtype.UUID
	Status     RefreshStatus
	Ip         *netip.Addr
	UserAgent  pgtype.Text
	ExpiresAt  pgtype.Timestamptz
	CreatedAt  pgtype.Timestamptz
	LastUsedAt pgtype.Timestamptz
}

type Role struct {
	ID          uuid.UUID
	Name        string
	Description string
}

type User struct {
	ID               uuid.UUID
	Email            string
	PasswordHash     string
	DisplayName      string
	Status           UserStatus
	MfaEnabled       bool
	MfaSecretEnc     []byte
	FailedLoginCount int32
	LockedUntil      pgtype.Timestamptz
	LastLoginAt      pgtype.Timestamptz
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
}

type UserRole struct {
	UserID    uuid.UUID
	RoleID    uuid.UUID
	GrantedBy pgtype.UUID
	GrantedAt pgtype.Timestamptz
}

type VerificationToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash []byte
	Purpose   TokenPurpose
	ExpiresAt pgtype.Timestamptz
	UsedAt    pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
}