# specs/03-api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Identity Hub API
  version: 0.1.0
  description: |
    Proveedor de identidad: alta administrativa, autenticación, emisión y rotación de
    tokens, segundo factor, autorización de aplicaciones, control por roles y auditoría.

    Cada operación declara `x-requirement` con el identificador del requisito que
    implementa (ver specs/01-requirements.md).

    El refresh token viaja exclusivamente en la cookie `refresh_token`, con
    `HttpOnly`, `Secure`, `SameSite=Strict` y `Path=/api/v1/auth`. Las operaciones
    de refresh y logout exigen esa cookie y usan `POST`; la renovación devuelve JSON
    y logout confirma con `204`. El resto de la API usa exclusivamente
    `Authorization: Bearer`, nunca cookies, por lo que no queda expuesta a CSRF.

    La API de la consola se entrega a través de Nginx y comparte origen. CORS solo se
    habilita para el origen configurado de Contabilidad en `POST /oauth/token` y en el
    JWKS, siempre sin credenciales. Los nombres OAuth en `snake_case` son la única
    excepción intencional a la convención `camelCase`: los define RFC 6749.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Desarrollo local (a través de Nginx)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [auth]
      summary: Aceptar una invitación y definir la contraseña
      description: Consume un token de invitación de un solo uso y activa la cuenta.
      x-requirement: RF-002
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvitationAcceptRequest' }
      responses:
        '204': { description: Invitación aceptada y cuenta activada }
        '400': { $ref: '#/components/responses/BadRequest' }
        '410':
          description: El token expiró o ya fue usado
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/v1/admin/users:
    post:
      operationId: createEmployee
      tags: [admin]
      summary: Dar de alta a un empleado y enviar su invitación
      x-requirement: RF-001
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateUserRequest' }
      responses:
        '201':
          description: Cuenta creada en estado pending_verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: El correo ya está registrado
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    get:
      operationId: listUsers
      tags: [admin]
      summary: Listar y buscar cuentas
      x-requirement: RF-010
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Página de cuentas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/users/{userId}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Consultar una cuenta
      x-requirement: RF-010
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Cuenta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Cambiar el estado o los roles de una cuenta
      description: Un administrador no puede deshabilitarse ni asignarse roles a sí mismo (invariantes 6 y 11).
      x-requirement: RF-010
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateUserRequest' }
      responses:
        '200':
          description: Cuenta actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token Ed25519. Validar contra el JWKS; se rechaza cualquier `alg` distinto de `EdDSA`.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Query:
      name: q
      in: query
      description: Búsqueda parcial por correo o nombre
      schema: { type: string, maxLength: 200 }
    Status:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/UserStatus' }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 500 }
  responses:
    BadRequest:
      description: Petición mal formada o que no supera la validación
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Credenciales ausentes, inválidas o expiradas
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Autenticado pero sin permiso para este recurso
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: El recurso no existe
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: Error en formato RFC 7807. Nunca revela si un correo está registrado.
      required: [type, title, status]
      properties:
        type:     { type: string, format: uri, example: 'https://identity.local/problems/invalid-credentials' }
        title:    { type: string, example: 'Credenciales inválidas' }
        status:   { type: integer, example: 401 }
        detail:   { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Detalle por campo en errores de validación
          items:
            type: object
            required: [field, message]
            properties:
              field:   { type: string }
              message: { type: string }
...
    UserStatus:
      type: string
      enum: [pending_verification, active, locked, disabled]
    Role:
      type: string
      enum: [admin, user, contabilidad.senior, contabilidad.analista]
    User:
      type: object
      required: [id, email, displayName, status, roles, mfaEnabled, createdAt]
      properties:
        id:          { type: string, format: uuid }
        email:       { type: string, format: email }
        displayName: { type: string }
        status:      { $ref: '#/components/schemas/UserStatus' }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        mfaEnabled:  { type: boolean, description: Siempre verdadero para cuentas activas; no requiere enrolamiento }
        createdAt:   { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
    UserPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/User' }
        nextCursor: { type: string, nullable: true }
    AdminCreateUserRequest:
      type: object
      required: [email, displayName, roles]
      properties:
        email:       { type: string, format: email, maxLength: 254 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/Role' }
    InvitationAcceptRequest:
      allOf:
        - $ref: '#/components/schemas/TokenRequest'
        - type: object
          required: [password]
          properties:
            password: { type: string, minLength: 12, maxLength: 128, format: password }
...
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
...
    AdminUpdateUserRequest:
      type: object
      properties:
        status: { $ref: '#/components/schemas/UserStatus' }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
...